requirements.txt
absl-py==0.9.0
astor==0.8.1
bayesian-optimization==1.1.0
bert-tensorflow==1.0.1
boto==2.49.0
boto3==1.12.15
botocore==1.15.15
cachetools==4.0.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
Click==7.0
cryptography==2.8
cycler==0.10.0
Cython==0.29.15
docutils==0.15.2
filelock==3.0.12
future==0.18.2
gast==0.2.2
gensim==3.8.1
google-auth==1.11.2
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
grpcio==1.27.2
h5py==2.10.0
Hello-world-app==0.0.0
idna==2.9
jmespath==0.9.5
joblib==0.14.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
langcodes==1.4.1
libsvm==3.23.0.3
marisa-trie==0.7.5
Markdown==3.2.1
matplotlib==3.1.3
msgpack==1.0.0
nltk==3.4.5
nose==1.3.7
numpy==1.18.1
oauthlib==3.1.0
opt-einsum==3.2.0
pandas==1.0.1
pexpect==4.8.0
protobuf==3.11.3
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
pyOpenSSL==19.1.0
pyparsing==2.4.6
python-dateutil==2.8.1
python-magic==0.4.15
pytz==2019.3
PyYAML==5.3
regex==2018.2.21
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
s3transfer==0.3.3
sacremoses==0.0.38
scikit-learn==0.22.1
scipy==1.4.1
seaborn==0.10.0
sentencepiece==0.1.85
six==1.14.0
smart-open==1.9.0
tensorboard==1.11.0
tensorflow==1.11.0
tensorflow-estimator==2.1.0
tensorflow-hub==0.7.0
termcolor==1.1.0
tokenizers==0.5.2
torch==1.4.0
tqdm==4.43.0
transformers==2.5.1
urllib3==1.25.8
Werkzeug==1.0.0
wordfreq==2.2.1
wrapt==1.12.0
